+++ packages/client/src/components/ForumPageCard/index.tsx
@@ -1,0 +1,1 @@
+export { ForumPageCard } from './ForumPageCard';
